refactor: 아이콘 적용
